renamed create_agent to agent
